rename drawable
