+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         </activity>
         <activity android:name=".MovieList.MovieListActivity"/>
         <activity android:name=".ZukanList.ZukanListActivity"/>
-        <activity android:name=".ZukanDetail.ZukanDetailMain"/>
+        <activity android:name=".ZukanDetail.ZukanDetaiActivity"/>
         <activity android:name=".panframe.SimplePlayerActivity"/>
         <activity android:name=".Quiz.QuizActivity"/>
     </application>
