+++ app/src/main/java/xyz/virtual_diving/projectmainver2/ZukanDetail/zukanDetailFragmentPagerAdapter.java
@@ -9,7 +9,7 @@ import android.support.v4.app.FragmentPagerAdapter;
  * Created by b1014100 on 2016/06/13.
  */
 public class zukanDetailFragmentPagerAdapter extends FragmentPagerAdapter {
-   public static ZukanDetail zukanDetail = ZukanDetailMain.zukanDetail;
+   public static ZukanDetail zukanDetail = ZukanDetaiActivity.zukanDetail;
 
     public zukanDetailFragmentPagerAdapter(FragmentManager fm) {
         super(fm);
