--- app/src/main/res/drawable/dmicon.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_pressed="true"
-        android:drawable="@drawable/new_diving_logo2"/>
-    <item android:state_pressed="false"
-        android:drawable="@drawable/new_diving_logo" />
-</selector>
