+++ app/src/main/res/drawable/homeactivity_dmicon.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_pressed="true"
+        android:drawable="@drawable/homeactivity_zukanlistlogo2"/>
+    <item android:state_pressed="false"
+        android:drawable="@drawable/homeactivity_zukanlistlogo" />
+</selector>
