+++ app/src/main/res/drawable/homeactivity_titlelogo.png
