+++ app/src/main/res/drawable/homeactivity_zukanlistlogo.png
