+++ app/src/main/res/drawable/homeactivity_zukanlistlogo2.png
