+++ app/src/main/res/drawable/zukandetail_action_forward.png
