+++ app/src/main/res/layout/movielist_listitem.xml
@@ -44,6 +44,6 @@
         android:layout_width="90dp"
         android:layout_height="90dp"
         android:layout_gravity="right"
-        android:src="@drawable/allow_right"
+        android:src="@drawable/movielist_allowright"
         />
 </LinearLayout>
